functions/requirements.txt
firebase_functions~=0.1.0
openai
anthropic
python-dotenv
numpy
sentence_transformers